pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "sonnylabs-mcp-security-proxy"
version = "0.1.0"
authors = [
  { name = "SonnyLabs" },
]
description = "MCP security proxy for detecting prompt injections using SonnyLabs API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.104.1",
  "pydantic>=2.4.2",
  "uvicorn>=0.23.2",
  "python-dotenv>=1.0.0",
  "httpx>=0.27",
  "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
